requirements.txt
Django>=3.0
djangorestframework
djangorestframework-jwt
requests
django-cors-headers
